scripts/codegen/requirements.txt
PyYAML>=6.0.2

Jinja2>=3.1.6

setuptools>=80.0.1
wheel>=0.45.1